--- src/components/GNBForm/Sys.js
@@ -1,763 +1,0 @@
-import styled from 'styled-components';
-import { AiOutlineStar, AiOutlineInfoCircle, AiOutlineSearch, AiOutlinePaperClip,  AiOutlineFolderOpen, AiOutlineFileText, AiFillStar } from 'react-icons/ai';
-import { MdOutlineRefresh } from 'react-icons/md';
-
-import axios from 'axios';
-import { useEffect, useState, useRef } from 'react';
-import { useSelector } from 'react-redux';
-
-export default function Sys(){
-  const [gnbList, setGnbList] = useState([]);
-  const [menuDetail, setMenuDetail] = useState([]);
-  const [result, setResult] = useState([]);
-  const [detail, setDetail] = useState([false, false]);
-  const [favor, setFavor] = useState(false);
-
-  const empId = useSelector(state => state.gnbMenu.key);
-  const menuId = 8;  // 현재 페이지 id 가져오기 (나중에)
-    useEffect(() => {
-      // gnb 리스트 가져오기
-      function GnbApi(){
-        axios({
-          url: `/setting/menu/gnb`,
-          method:'get',
-          baseURL : 'http://localhost:8080',
-        }).then(res => {setGnbList(res.data.data)});
-      }
-
-      function Favor(){
-        axios({
-          url: `/setting/favor?empId=${empId}&menuId=${menuId}`,
-          method:'get',
-          baseURL : 'http://localhost:8080',
-        }).then(res => {console.log(res.data);setFavor(res.data)});
-      }
-
-      GnbApi();
-      Favor();
-    }, [empId, menuId])
-
-    // 대메뉴/메뉴 디테일 on/off
-    function menuDetailHandler(type, menuDetail){
-      if (type === 'gnbDetail') {
-        setDetail([true, false]);
-      }
-      if (type === 'menuDetail') {
-        setDetail([false, true]);
-      }
-      if (type === 'newGnb') {
-        setDetail([true, false]);
-      }
-      if (type === 'newMenu') {
-        setDetail([false, true]);
-      }
-      setMenuDetail(menuDetail);
-      console.log(type, menuDetail);
-    }
-
-    function detailOff(){
-      setDetail([false, false])
-    }
-
-    // 메뉴 검색 결과
-    function searchAPI(event){
-      console.log('searchAPI');
-      event.preventDefault();
-      const formData = new FormData(document.getElementById('searchForm'));
-//      console.log(formData);
-//      console.log(formData.get("gnbName"), formData.get("name"));
-      
-      axios({
-        url: `/setting/menu/search?gnbName=${formData.get("gnbName")}%25&name=%25${formData.get("name")}%25`,
-        method:'get',
-        baseURL : 'http://localhost:8080',
-        data: {
-          formData
-        }
-      }).then(res => {
-        console.log('setResult : ', res.data.data);
-        setResult(res.data.data);
-      });
-    
-//      console.log(formData);
-    }
-
-    // 즐겨찾기 추가/삭제 요청
-    function FavorApi(){
-      console.log(favor);
-      if (favor) {
-        // 즐겨찾기가 되어 있으면 삭제 요청
-        axios({
-          url: `/setting/favor?empId=${empId}&menuId=${menuId}`,
-          method:'delete',
-          baseURL : 'http://localhost:8080'
-      }).then(res => {
-        console.log('favor off : ', res.data);
-        setFavor(!favor);
-      });
-    } else {
-      // 즐겨찾기가 안되어 있으면 즐겨찾기 추가 요청
-      axios({
-        url: `/setting/favor?empId=${empId}&menuId=${menuId}`,
-        method:'post',
-        baseURL : 'http://localhost:8080',
-        data : {
-          empId: empId,
-          menuId: menuId
-        }
-      }).then(res => {
-        console.log('favor on : ', res.data);
-        setFavor(!favor);
-      });
-    }
-  }
-  
-    return (
-        <Module>
-            <Nav>
-              <div id ='title'>메뉴설정</div>
-              <div id ='btn'>
-                <button>변경이력</button>
-                <button onClick={() => {menuDetailHandler('newGnb', '')}}>대메뉴추가</button>
-                <button onClick={() => {menuDetailHandler('newMenu', '')}}>메뉴추가</button>
-                <hr />
-                <div onClick={FavorApi}>{favor ? <AiFillStar /> : <AiOutlineStar/>}</div>
-              </div>
-            </Nav>
-            <Info>
-              <div>
-              <AiOutlineInfoCircle />&nbsp; info
-              </div>
-            </Info>
-            <Form>
-                <MenuList value={gnbList} api={menuDetailHandler}/>
-                <MenuTree>
-                    <SearchTree id='searchForm'>
-                        <select name="gnbName">
-                          {
-                            gnbList.map((a, i) => (
-                              <option key={i} id='gnbName' value={a['name']}>{a['name']}</option>
-                            ))
-                          }
-                          </select>
-                        <select><option>전체</option></select>
-                        <input id='name' placeholder=' 메뉴명 검색' name='name'/>
-                        <AiOutlineSearch onClick={(e) => {searchAPI(e)}}/>
-                        <br />
-                        <div>
-                        <span>메뉴목록</span>
-                        <select name="filter">
-                          <option value="all">필터</option>
-                          <option value='admin'>관리자 메뉴</option>
-                          <option value='general'>사용자메뉴</option>
-                        </select>
-                        </div>
-                    </SearchTree>
-                    <div>{result.map((a, i) => (
-                      <div onClick={() => {setDetail([false, true]);setMenuDetail(a)}}>{a['name']}</div>
-                    ))}
-                    </div>
-                </MenuTree>
-                { detail[0] && <MenuDetail value={menuDetail} api={detailOff} on={detail[0]}/>}
-                { detail[1] && <NewMenu value={menuDetail} api={detailOff} on={detail[1]}/>}
-                
-            </Form>
-        </Module>
-    )
-}
-
-export function TreeItem(props){
-  console.log('treeitem : ', props.value, props.parId);
-  return (
-    <div>{props.value.map((a, i) => (
-      props.parId===a['parId'] && a['id'] !== props.parId ? 
-      <Item>{a['sub'] === 0 ? <AiOutlineFolderOpen /> : <AiOutlineFileText /> } {a['name']}
-        <TreeItem value={props.value} parId={a['id']}/>
-      </Item> : null
-    ))}
-    </div>
-  )
-
-}
-  
-export function MenuList(props) {
-console.log(props.value[0]);
-    return (
-      <ListBox>
-        <div>
-        {
-          props.value.map((a, i) => (
-            <MenuItem key={i} onClick={()=>{props.api('gnbDetail', a)}}>
-            <img src={a['iconUrl']} alt='i'/><div id='menuitemname'>{a['name']}</div>
-            </MenuItem>
-          ))
-        }
-        </div>
-      </ListBox>
-    );
-  }
-
-export function MenuDetail(props) {
-  const [isDragging, setIsDragging] = useState(false);
-  const [data, setData] = useState([]);
-  const [iconFile, setIconFile] = useState([]);
-  const [iconKey, setIconKey] = useState([]);
-  const nameRef = useRef(null);
-  const sortRef = useRef(null);
-
-  
-  const onDragEnter = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-  const onDragLeave = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-  const onDragOver = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (e.dataTransfer.files) {
-      setIsDragging(true);
-    }
-  };
-  const onDrop = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIconFile(e.dataTransfer.files[0]);
-    console.log(e.dataTransfer.files[0]['name']);
-    setIconKey(e.dataTransfer.files[0]['name']);
-    setIsDragging(false);
-  };
-      const loadData = async (e) => {
-        e.preventDefault();
-
-        // 이미지 데이터 보내기
-        try {
-            await axios({
-                url: `/setting/menu/img`,
-                method:'post',
-                baseURL : 'http://localhost:8080',
-                headers: { 'Content-Type': 'multipart/form-data' },
-                data:{
-                    //id : id,
-                    iconFile:  iconFile
-                }
-            });
-            console.log(iconFile);
-            console.log('successed sending img...');
-        } catch (error) {
-            console.log('failed sending img...');
-        }
-        
-        // form 데이터 보내기 
-        try {
-            await axios({
-                url: `/setting/menu`,
-                method:'post',
-                baseURL : 'http://localhost:8080',
-                data:{
-                    id:  data['id'],
-                    name:  nameRef.current.value,
-                    enabledYN:  e.target.enabledYN.value,
-                    sortOrder:  sortRef.current.value
-                }
-            });
-            console.log('successed sending Data...');
-        } catch (error) {
-            console.log('failed sending Data...');
-        }
-    }
-
-    useEffect (() => {
-    if (props.value === undefined){
-     setData(JSON.parse('[{"name":"", "sortOrder":""}]'));
-    } else {
-    setData(props.value);
-    }
-    }, [props.value]);
-
-  return (
-    props.on &&
-    <DetailDiv>
-      
-      <form onSubmit={loadData}>
-        <div style={{display: 'flex', justifyContent:'space-between'}}> 
-      <span>•대메뉴 정보</span><div style={{display: 'flex'}}><button type='submit'>저장</button> <hr /> <span onClick={props.api}>X</span></div>
-      </div>
-      <table>
-        <tbody>
-          <tr><td>메뉴명</td><td><input type='text' id='name' placeholder={data['name']} ref={nameRef}/></td></tr>
-          <tr><td>사용여부</td><td id='enabledYN'><div><input className='radio' type='radio' value='1' name='enabledYN'/>사용<input className='radio'  type='radio' value='0'  name='enabledYN' />미사용</div></td></tr>
-          <tr><td>정렬</td><td><input type='number' id='sortOrder' placeholder={data['sortOrder']} ref={sortRef}/></td></tr>
-          <tr><td>아이콘(48*44)<MdOutlineRefresh /></td>
-          <td style={{display: 'block'}}>
-            <div style={{display: 'flex' }}>
-              <div style={{border: '1px solid rgb(171,172,178)', width:'calc(100% - 25px)', height:'25px', color:'rgb(171,172,178)', fontWeight: 'bold', fontSize: 'small', paddingTop: '4px'}}> &nbsp;마우스로 파일을 끌어오세요</div>
-            <input type='file' id="iconFile" onChange={(e) => {setIconFile(e.target.files[0])}} key={iconKey}/>
-            <AiOutlinePaperClip for='iconFile' className='iconFileInput' />
-            </div>
-          <hr/>
-          <div id="DnDBox" onDragEnter={onDragEnter} onDragLeave={onDragLeave} onDragOver={onDragOver} onDrop={onDrop} isDragging={isDragging}>
-            <IconImageList />
-          </div>
-          </td></tr>
-        </tbody>
-      </table>
-      
-      </form>
-    </DetailDiv>
-  )
-
-}
-
-export function IconImageList(){
-  const [icon, setIcon] = useState([]);
-//  console.log(icon);
-useEffect(() => {
-  axios({
-    url: `/setting/menu/iconList`,
-    method:'get',
-    baseURL : 'http://localhost:8080',
-  }).then(res => {
-    console.log(res.data);
-    setIcon(res.data);
-  })
-}, [])
-
-  return(
-    <div>
-      {
-      icon.map((a, i) => (
-        <img src={a} alt='i' key={i}/>
-      ))
-      }      
-    </div>
-  )
-}
-
-export function NewGnb(props) {
-  const [isDragging, setIsDragging] = useState(false);
-  const [data, setData] = useState([]);
-  const [iconFile, setIconFile] = useState([]);
-  const [iconKey, setIconKey] = useState([]);
-  const nameRef = useRef(null);
-  const sortRef = useRef(null);
-
-  
-  const onDragEnter = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-  const onDragLeave = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-  const onDragOver = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (e.dataTransfer.files) {
-      setIsDragging(true);
-    }
-  };
-  const onDrop = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIconFile(e.dataTransfer.files[0]);
-    console.log(e.dataTransfer.files[0]['name']);
-    setIconKey(e.dataTransfer.files[0]['name']);
-    setIsDragging(false);
-  };
-      const loadData = async (e) => {
-        e.preventDefault();
-
-        // 새로운 아이디로 추가 요청을 보내야 함.
-        // 이미지 데이터 보내기
-        try {
-            await axios({
-                url: `/setting/menu/img`,
-                method:'post',
-                baseURL : 'http://localhost:8080',
-                headers: { 'Content-Type': 'multipart/form-data' },
-                data:{
-                    //id : id,
-                    iconFile:  iconFile
-                }
-            });
-            console.log(iconFile);
-            console.log('successed sending img...');
-        } catch (error) {
-            console.log('failed sending img...');
-        }
-        
-        // form 데이터 보내기 
-        try {
-            await axios({
-                url: `/setting/menu`,
-                method:'post',
-                baseURL : 'http://localhost:8080',
-                data:{
-                    id:  data['id'],
-                    name:  nameRef.current.value,
-                    enabledYN:  e.target.enabledYN.value,
-                    sortOrder:  sortRef.current.value
-                }
-            });
-            console.log('successed sending Data...');
-        } catch (error) {
-            console.log('failed sending Data...');
-        }
-    }
-
-    useEffect (() => {
-    if (props.value === undefined){
-     setData(JSON.parse('[{"name":"", "sortOrder":""}]'));
-    } else {
-    setData(props.value);
-    }
-    }, [props.value]);
-
-  return (
-    props.on &&
-    <DetailDiv>
-      
-      <form onSubmit={loadData}>
-        <div style={{display: 'flex', justifyContent:'space-between'}}> 
-      <span>•대메뉴 정보</span><div style={{display: 'flex'}}><button type='submit'>저장</button> <hr /> <span onClick={props.api}>X</span></div>
-      </div>
-      <table>
-        <tbody>
-          <tr><td>메뉴명</td><td><input type='text' ref={nameRef}/></td></tr>
-          <tr><td>사용여부</td><td id='enabledYN'><div><input className='radio' type='radio' value='1' name='enabledYN'/>사용<input className='radio'  type='radio' value='0'  name='enabledYN' />미사용</div></td></tr>
-          <tr><td>정렬</td><td><input type='number' id='sortOrder' placeholder={data['sortOrder']} ref={sortRef}/></td></tr>
-          <tr><td>아이콘(48*44)<MdOutlineRefresh /></td>
-          <td style={{display: 'block'}}>
-            <div style={{display: 'flex' }}>
-              <div style={{border: '1px solid rgb(171,172,178)', width:'calc(100% - 25px)', height:'25px', color:'rgb(171,172,178)', fontWeight: 'bold', fontSize: 'small', paddingTop: '4px'}}> &nbsp;마우스로 파일을 끌어오세요</div>
-            <input type='file' id="iconFile" onChange={(e) => {setIconFile(e.target.files[0])}} key={iconKey}/>
-            <AiOutlinePaperClip for='iconFile' className='iconFileInput' />
-            </div>
-          <hr/>
-          <div id="DnDBox" onDragEnter={onDragEnter} onDragLeave={onDragLeave} onDragOver={onDragOver} onDrop={onDrop} isDragging={isDragging}>
-            <IconImageList />
-          </div>
-          </td></tr>
-        </tbody>
-      </table>
-      
-      </form>
-    </DetailDiv>
-  )
-
-}
-
-export function NewMenu(props) {
-  const [data, setData] = useState([]);
-  const [iconFile, setIconFile] = useState([]);
-  const nameRef = useRef(null);
-  const sortRef = useRef(null);
-
-  // 메뉴 정보 추가로 수정.
-      const loadData = async (e) => {
-        e.preventDefault();
-
-        // 새로운 아이디로 추가 요청을 보내야 함.
-        // 이미지 데이터 보내기
-        try {
-            await axios({
-                url: `/setting/menu/img`,
-                method:'post',
-                baseURL : 'http://localhost:8080',
-                headers: { 'Content-Type': 'multipart/form-data' },
-                data:{
-                    //id : id,
-                    iconFile:  iconFile
-                }
-            });
-            console.log(iconFile);
-            console.log('successed sending img...');
-        } catch (error) {
-            console.log('failed sending img...');
-        }
-        
-        // form 데이터 보내기 
-        try {
-            await axios({
-                url: `/setting/menu`,
-                method:'post',
-                baseURL : 'http://localhost:8080',
-                data:{
-                    id:  data['id'],
-                    name:  nameRef.current.value,
-                    enabledYN:  e.target.enabledYN.value,
-                    sortOrder:  sortRef.current.value
-                }
-            });
-            console.log('successed sending Data...');
-        } catch (error) {
-            console.log('failed sending Data...');
-        }
-    }
-
-  return (
-    props.on &&
-    <DetailDiv>
-      
-      <form onSubmit={loadData}>
-        <div style={{display: 'flex', justifyContent:'space-between'}}> 
-      <span>•메뉴 정보</span><div style={{display: 'flex'}}><button type='submit'>저장</button> <hr /> <span onClick={props.api}>X</span></div>
-      </div>
-      <table>
-        <tbody>
-          <tr><td>상위메뉴</td><td><select><option>알림설정</option></select></td></tr>
-          <tr><td>메뉴명</td><td><input type='text' ref={nameRef} placeholder={props.value['name']}/></td></tr>
-          <tr><td>사용여부</td><td id='enabledYN'><div><input className='radio' type='radio' value='1' name='enabledYN'/>사용<input className='radio'  type='radio' value='0'  name='enabledYN' />미사용</div></td></tr>
-          <tr><td>정렬</td><td><input type='number' id='sortOrder' placeholder={props.value['sortOrder']} ref={sortRef}/></td></tr>
-        </tbody>
-      </table>
-      
-      </form>
-    </DetailDiv>
-  )
-
-}
-
-export const Module = styled.div`
-background-color: white;
-border: 1px solid rgb(171,172,178);
-width:95%;
-height: 95%;
-`
-export const Nav = styled.div`
-border-top: 5px solid rgb(20,136,247);
-border-bottom: 1px solid grey;
-display: flex;
-color: black;
-width: 100%;
-justify-content: space-between;
-
-> #title {
-  margin: 10px;
-  margin-left: 20px;
-  font-size: large;
-  font-weight: bold;
-  color: rgb(32,35,44);
-}
-> #btn {
-  display: flex;
-  margin-top: 5px;
-  > * {
-    
-    height: 30px;
-    margin: 5px;
-  }
-  > svg {
-    margin-top: 10px;
-    width: 20px;
-    height: 20px;
-    color: gray;
-  }
-}
-`;
-export const Info = styled.div`
-display: flex;
-justify-content: center;
-
-> div {
-  margin: 10px;
-  padding: 10px;
-  padding-left:15px;
-  width: 100%;
-  background-color: rgb(214,236,248);
-  border: 1px solid rgb(146,183,214);
-  border-radius: 5px;
-
-  color: black;
-  height: 20px;
-  font-weight: bold;
-}
-`;
-export const Form = styled.div`
-display: flex;
-justify-content: flex-start;
-width: 100%;
-height: 100%;
-margin-bottom: 10px;
-`;
-export const MenuTree = styled.div`
-margin:10px;
-height:75%;
-background-color: white;
-width: 400px;
-border: 1px solid rgb(171,172,178);
-color: black;
-`;
-export const SearchTree = styled.form`
-padding: 5px;
-height: 100px;
-background-color: rgb(240,245,248);
-border-bottom: 1px solid rgb(171,172,178);
-color: black;
-> select {
-  width:170px;
-  height:25px;
-  margin: 5px;
-  margin-left: 10px;
-}
-> input {
-  width: 320px;
-  height: 20px;
-  margin: 5px;
-  margin-left: 10px;
-}
-> svg {
-  width: 20px;
-  height: 20px;
-}
-
-> div {
-  display: flex;
-  justify-content: space-between;
-  font-weight: bold;
-  margin: 5px;
-
-  > span {
-    font-size: small;
-  }
-  > select {
-    font-weight: bold;
-    position: relative;
-    border: none;
-    background-color: rgb(240,245,248);
-    text-align: right;
-    > option {
-      background-color: white;
-    }
-  }
-}
-`
-export const Item = styled.div`
-margin-left: 10px;
-margin-top: 5px;
-`;
-export const DetailDiv = styled.div`
-margin:5px;
-color: black;
-weight: 610px;
-height: 550px;
-> form > div {
-
-> span {
-  font-weight: bold;
-  margin-bottom: 10px;
-  }
-  > div > * {
-    margin: 5px;
-  }
-}
-> form > table {
-  border-top: 2px solid black;
-  border-collapse: collapse;
-  > tbody{
-    > tr {  
-      border-bottom: 1px solid rgb(171,172,178);
-        > td:nth-child(1) {
-          background-color: rgb(240,245,248);
-        font-weight: bold;
-        text-align: right;
-        padding-right: 10px;
-        width:150px;
-        }
-        > td:nth-child(2) {
-          width:400px;
-          margin: 10px;
-          display: flex;
-          > input {
-            width: 100%;
-            height: 25px;
-          }   
-          > div {
-            > #iconFile {
-              display: none;
-            }
-            > .iconFileInput {
-              width: 25px;
-              height: 25px;
-              margin: 5px;
-            }
-  
-          }  
-
-        }
-      }
-      > tr > td > #DnDBox {
-        width: 400px;
-        height: 300px;
-        
-        > div {
-          display: flex;
-          flex-direction: row;
-          flex-wrap: wrap;
-          > img {
-            width: 40px;
-            height: 40px;
-            padding : 5px;
-            margin: 5px;
-            border: 1px solid rgb(171,172,178);
-          }
-        }
-      }
-  }
-}
-
-`;
-export const ListBox = styled.div`
-display: flex;
-justify-content: center;
-background-color: rgb(240,245,248);
-border: 1px solid rgb(171,172,178);
-border-top: 3px solid gray;
-margin: 10px;
-paddig: 10px;
-width: 250px;
-height: 75%;
-> div {
-  display: block;
-  
-  width: 220px;
-  height: 90%;
-  overflow: scroll;
-  &::-webkit-scrollbar{
-      display:none;
-  }  
-  > div > img {
-    margin: 7px;
-    width: 40px;
-    height: 40px;
-  }
-}
-`;
-  export const MenuItem = styled.div`
-margin-top:5px;
-
-display: flex;
-justify-content: flex-start;
-
-height: 50px;
-
-background-color: white;
-border: 1px solid rgb(171,172,178);
-color: black;
-
-padding: 10px;
-
-text-align: center;
-font-size: medium;
-> #menuitemname {
-  padding-top: 5%;
-}
-`;
